convert title to trait
