+++ app/controllers/app.php
@@ -20,30 +20,6 @@ class App extends Controller
 	}
 
 	/**
-	 * Returns the title of the current page
-	 * @return string
-	 */
-	public function title()
-	{
-		if (is_home()) {
-			if ($home = get_option('page_for_posts', true)) {
-				return get_the_title($home);
-			}
-			return __('Latest Posts', 'sage');
-		}
-		if (is_archive()) {
-			return get_the_archive_title();
-		}
-		if (is_search()) {
-			return sprintf(__('Search Results for %s', 'sage'), get_search_query());
-		}
-		if (is_404()) {
-			return __('Not Found', 'sage');
-		}
-		return get_the_title();
-	}
-
-	/**
 	 * Returns array of comments for a post
 	 * @return array
 	 */
@@ -53,7 +29,7 @@ class App extends Controller
 		$comments = $comments->getComments();
 		return $comments;
 	}
-
+	
 	/**
 	 * Returns cookie notice from Advanced Custom Fields
 	 * @return array
